+++ src/dimensions/correlationdim.jl
@@ -1,14 +1,16 @@
+import ProgressMeter
 #######################################################################################
 # Original correlation sum
 #######################################################################################
 using Distances, Roots
 export correlationsum, grassberger_dim, boxed_correlationsum,
-estimate_r0_buenoorovio, data_boxing
+estimate_r0_buenoorovio, data_boxing, autoprismdim, estimate_r0_theiler
 
 """
     correlationsum(X, ε::Real; w = 0, norm = Euclidean(), q = 2) → C_q(ε)
 Calculate the `q`-order correlation sum of `X` (`Dataset` or timeseries)
-for a given radius `ε` and `norm`.
+for a given radius `ε` and `norm`. They keyword `show_progress = false` can be used
+to display a progress bar for large `X`.
 
 The function [`boxed_correlationsum`](@ref) is faster and should be preferred over this one.
 
@@ -34,40 +36,44 @@ explicit definition of the q-order correlationsum.
     correlationsum(X, εs::AbstractVector; w, norm, q) → C_q(ε)
 
 If `εs` is a vector, `C_q` is calculated for each `ε ∈ εs`.
-If also `q=2`, some strong optimizations are done, but this requires the allocation
-a matrix of size `N×N`. If this is larger than your available memory please use instead:
-```julia
-[correlationsum(..., ε) for ε in εs]
-```
+If also `q=2`, we attempt to do further optimizations are done, if the allocation
+a matrix of size `N×N` is possible
 
-See [`grassberger`](@ref) for more.
-See also [`takens_best_estimate`](@ref).
+See [`grassberger`](@ref) for more. See also [`takens_best_estimate`](@ref).
 
 [^Grassberger]: Peter Grassberger (2007) [Grassberger-Procaccia algorithm. Scholarpedia, 2(5):3043.](http://dx.doi.org/10.4249/scholarpedia.3043)
 
 [^Kantz]: Kantz, H., & Schreiber, T. (2003). [More about invariant quantities. In Nonlinear Time Series Analysis (pp. 197-233). Cambridge: Cambridge University Press.](https://doi:10.1017/CBO9780511755798.013)
 """
-function correlationsum(X, ε; q = 2, norm = Euclidean(), w = 0)
+function correlationsum(X, ε; q = 2, norm = Euclidean(), w = 0, show_progress = false)
     if q == 2
-        correlationsum_2(X, ε, norm, w)
+        correlationsum_2(X, ε, norm, w, show_progress)
     else
-        correlationsum_q(X, ε, q, norm, w)
+        correlationsum_q(X, ε, q, norm, w, show_progress)
     end
 end
 
-function correlationsum_2(X, ε::Real, norm = Euclidean(), w = 0)
-    N, C = length(X), zero(eltype(X))
-    for (i, x) in enumerate(X)
+function correlationsum_2(X, ε::Real, norm, w, show_progress)
+    N = length(X)
+    if show_progress
+        progress = ProgressMeter.Progress(N; desc = "Correlation sum: ", dt = 1.0)
+    end
+    C = zero(eltype(X))
+    @inbounds for (i, x) in enumerate(X)
         for j in i+1+w:N
             C += evaluate(norm, x, X[j]) < ε
         end
+        show_progress && ProgressMeter.next!(progress)
     end
     return C * 2 / ((N-w-1)*(N-w))
 end
 
-function correlationsum_q(X, ε::Real, q, norm = Euclidean(), w = 0)
+function correlationsum_q(X, ε::Real, q, norm, w, show_progress)
     N, C = length(X), zero(eltype(X))
     normalisation = (N-2w)*(N-2w-one(eltype(X)))^(q-1)
+    if show_progress
+        progress = ProgressMeter.Progress(length(1+w:N-w); desc = "Correlation sum: ", dt = 1.0)
+    end
     for i in 1+w:N-w
         x = X[i]
         C_current = zero(eltype(X))
@@ -80,44 +86,65 @@ function correlationsum_q(X, ε::Real, q, norm = Euclidean(), w = 0)
             C_current += evaluate(norm, x, X[j]) < ε
         end
         C += C_current^(q - 1)
+        show_progress && ProgressMeter.next!(progress)
     end
     return (C / normalisation) ^ (1 / (q-1))
 end
 
 # Optimized version
-function correlationsum_2(X, εs::AbstractVector, norm = Euclidean(), w = 0)
+function correlationsum_2(X, εs::AbstractVector, norm, w, show_progress)
     @assert issorted(εs) "Sorted εs required for optimized version."
-    d = distancematrix(X, norm)
+    d = try
+        distancematrix(X, norm)
+    catch err
+        @warn "Couldn't create distance matrix ($(typeof(err))). Using slower algorithm..."
+        return [correlationsum_2(X, ε, norm, w, show_progress) for ε in εs]
+    end
+    return correlationsum_2_fb(X, εs, d, w, show_progress) # function barrier
+end
+function correlationsum_2_fb(X, εs, d, w, show_progress)
     Cs = zeros(eltype(X), length(εs))
     N = length(X)
     factor = 2/((N-w)*(N-1-w))
+    if show_progress
+        K = length(length(εs)÷2:-1:1)
+        M = K + length((length(εs)÷2 + 1):length(εs))
+        progress = ProgressMeter.Progress(M; desc = "Correlation sum: ", dt = 1.0)
+    end
+
     # First loop: mid-way ε until lower saturation point (C=0)
-    for k in length(εs)÷2:-1:1
+    for (ki, k) in enumerate(length(εs)÷2:-1:1)
         ε = εs[k]
         for i in 1:N
             @inbounds Cs[k] += count(d[j, i] < ε for j in i+1+w:N)
         end
+        show_progress && ProgressMeter.update!(progress, ki)
         Cs[k] == 0 && break
     end
     # Second loop: mid-way ε until higher saturation point (C=max)
-    for k in (length(εs)÷2 + 1):length(εs)
+    for (ki, k) in enumerate((length(εs)÷2 + 1):length(εs))
         ε = εs[k]
         for i in 1:N
             @inbounds Cs[k] += count(d[j, i] < ε for j in i+1+w:N)
         end
+        show_progress && ProgressMeter.update!(progress, ki+K)
         if Cs[k] ≈ 1/factor
             Cs[k:end] .= 1/factor
             break
         end
     end
+    show_progress && ProgressMeter.finish!(progress)
     return Cs .* factor
 end
 
-function correlationsum_q(X, εs::AbstractVector, q, norm = Euclidean(), w = 0)
+function correlationsum_q(X, εs::AbstractVector, q, norm, w, show_progress)
     @assert issorted(εs) "Sorted εs required for optimized version."
     Nε, T, N = length(εs), eltype(X), length(X)
     Cs = zeros(T, Nε)
     normalisation = (N-2w)*(N-2w-one(T))^(q-1)
+    if show_progress
+        progress = ProgressMeter.Progress(length(1+w:N-w); desc = "Correlation sum: ", dt = 1.0)
+    end
     for i in 1+w:N-w
         x = X[i]
         C_current = zeros(T, Nε)
@@ -133,6 +160,7 @@ function correlationsum_q(X, εs::AbstractVector, q, norm = Euclidean(), w = 0)
             end
         end
         Cs .+= C_current .^ (q-1)
+        show_progress && ProgressMeter.next!(progress)
     end
     return (Cs ./ normalisation) .^ (1/(q-1))
 end
@@ -174,72 +202,95 @@ function grassberger_dim(data::AbstractDataset, εs = estimate_boxsizes(data); k
 end
 
 ################################################################################
-# Correlationsum, but we distributed data to boxes beforehand
+# Boxed Correlation sum (we distribute data to boxes beforehand)
 ################################################################################
 """
-    boxed_correlationsum(data, εs, r0 = maximum(εs); q = 2 , M = size(data, 2), w = 0)
-    boxed_correlationsum(data; q = 2 , M = size(data, 2), w = 0)
+    boxed_correlationsum(X::Dataset, εs, r0 = maximum(εs); kwargs...) → Cs
 
-Estimate the box assisted q-order correlation sum[^Kantz2003] out of a
-dataset `data` for radii `εs` by splitting the data into boxes of size `r0`
-beforehand.
+Estimate the box assisted q-order correlation sum[^Kantz2003] `Cs` out of a
+dataset `X` for each radius in `εs`, by splitting the data into boxes of size `r0`
+beforehand. This method is much faster than [`correlationsum`](@ref), **provided that** the 
+box size `r0` is significantly smaller than then the attractor length.
+A good estimate for `r0` is [`estimate_r0_buenoorovio`](@ref).
+
+    boxed_correlationsum(X; kwargs...) → εs, Cs
+
+In this method the minimum inter-point distance and [`estimate_r0_buenoorovio`](@ref)
+are used to estimate good `εs` for the calculation, which are also returned.
+
+## Keywords
+* `q = 2`
+* `P = autoprismdim(data)` : The prism dimension.
+* `w = 0` : The [Theiler window](@ref).
+* `show_progress = false` : Whether to display a progress bar for the calculation.
 
 ## Description
 `C_q(ε)` is calculated for every `ε ∈ εs` and each of the boxes to then be
-summed up afterwards. If `M` is unequal to the dimension of the data, only the
-first `M` dimensions are considered for the box distribution (this is called the
-prism-assisted version).
-The method of splitting the data into boxes was
-implemented according to Theiler[^Theiler1987]. If only a dataset is given the
-radii `εs` and boxsize `r0` are chosen by calculating
-[`estimate_r0_buenoorovio`](@ref).
+summed up afterwards. The method of splitting the data into boxes was 
+implemented according to Theiler[^Theiler1987]. `w` is the [Theiler window](@ref).
+`P` is the prism dimension. If `P` is unequal to the dimension of the data, only the
+first `P` dimensions are considered for the box distribution (this is called the
+prism-assisted version). By default `P` is choosen automatically.
 
-`w` is the [Theiler window](@ref).
-
-The function is explicitly optimized for `q = 2` and becomes quite slow for `q ≠ 2`.
+The function is explicitly optimized for `q = 2` but becomes quite slow for `q ≠ 2`.
 
 See [`correlationsum`](@ref) for the definition of `C_q`
-and also [`data_boxing`](@ref).
+and also [`data_boxing`](@ref) to use the algorithm that splits data into boxes.
 
 [^Kantz]: Kantz, H., & Schreiber, T. (2003). [More about invariant quantities. In Nonlinear Time Series Analysis (pp. 197-233). Cambridge: Cambridge University Press.](https://doi:10.1017/CBO9780511755798.013)
 
 [^Theiler1987]: Theiler, [Efficient algorithm for estimating the correlation dimension from a set of discrete points. Physical Review A, 36](https://doi.org/10.1103/PhysRevA.36.4456)
 """
-function boxed_correlationsum(data; kwargs...)
-    r0 = estimate_r0_buenoorovio(data, M)
-    ε0 = minimum_pairwise_distance(data)[1]
-    εs = 10 .^ range(log10(ε0), log10(r0), length = 16)
-    boxed_correlationsum(data, εs, r0; kwargs...)
+function boxed_correlationsum(data; q = 2, P = autoprismdim(data), w = 0, show_progress=false)
+    r0, ε0 = estimate_r0_buenoorovio(data, P)
+    @assert  r0 < ε0 "The calculated box size was smaller than the minimum interpoint " *
+    "distance. Please choose manually."
+    εs = 10.0 .^ range(log10(ε0), log10(r0), length = 16)
+    boxed_correlationsum(data, εs, r0; q, P, w, show_progress)
 end
 
 function boxed_correlationsum(
         data, εs, r0 = maximum(εs);
-        q = 2, M = _autoprismdim(data), w = 0
+        q = 2, P = autoprismdim(data), w = 0,
+        show_progress = false,
     )
-    @assert M ≤ size(data, 2) "Prism dimension has to be lower or equal than " *
+    @assert P ≤ size(data, 2) "Prism dimension has to be lower or equal than " *
     "data dimension."
-    boxes, contents = data_boxing(data, r0, M)
+    boxes, contents = data_boxing(data, r0, P)
     if q == 2
-        boxed_correlationsum_2(boxes, contents, data, εs; w)
+        boxed_correlationsum_2(boxes, contents, data, εs; w, show_progress)
     else
-        boxed_correlationsum_q(boxes, contents, data, εs, q; w)
+        boxed_correlationsum_q(boxes, contents, data, εs, q; w, show_progress)
     end
 end
 
-function _autoprismdim(data)
+"""
+    autoprismdim(data, version = :bueno)
+
+An algorithm to find the ideal choice of a prism dimension for [`boxed_correlationsum`](@ref).
+`version = :bueno` uses `P=2`, while `version = :theiler` uses Theiler's original suggestion.
+"""
+function autoprismdim(data, version = :bueno)
     D = dimension(data)
-    L = length(data)
-    # The best choice given by Theiler is $P = 0.5\log_2N$ and should be used when $D$ exceeds $0.75\log_2N$.
-    return size(data, 2)
+    N = length(data)
+    if version == :bueno
+        return min(D, 2)
+    elseif version == :theiler
+        if D > 0.75 * log2(N)
+            return max(2, ceil(0.5 * log2(N)))
+        else
+            return D
+        end
+    end
 end
 
 """
-    data_boxing(data, r0, M = size(data, 2))
+    data_boxing(data, r0, P) → boxes, contents
 Distribute the `data` points into boxes of size `r0`. Return box positions
 and the contents of each box as two separate vectors. Implemented according to
 the paper by Theiler[^Theiler1987] improving the algorithm by Grassberger and
-Procaccia[^Grassberger1983]. If `M` is smaller than the dimension of the data,
-only the first `M` dimensions are considered for the distribution into boxes.
+Procaccia[^Grassberger1983]. If `P` is smaller than the dimension of the data,
+only the first `P` dimensions are considered for the distribution into boxes.
 
 See also: [`boxed_correlationsum`](@ref).
 
@@ -247,13 +298,12 @@ See also: [`boxed_correlationsum`](@ref).
 
 [^Grassberger1983]: Grassberger and Proccacia, [Characterization of strange attractors, PRL 50 (1983)](https://journals-aps-org.e-bis.mpimet.mpg.de/prl/abstract/10.1103/PhysRevLett.50.346)
 """
-function data_boxing(data, r0, M = size(data, 2))
-    @assert M ≤ size(data, 2) "Prism dimension has to be lower or equal than "*
-    "data dimension."
-    mini = minima(data)[1:M]
+function data_boxing(data, r0, P)
+    @assert P ≤ size(data, 2) "Prism dimension has to be ≤ than data dimension."
+    mini = minima(data)[1:P]
 
     # Map each datapoint to its bin edge and sort the resulting list:
-    bins = map(point -> floor.(Int, (point[1:M] - mini)/r0), data)
+    bins = map(point -> floor.(Int, (point[1:P] - mini)/r0), data)
     permutations = sortperm(bins, alg=QuickSort)
 
     boxes = unique(bins[permutations])
@@ -263,7 +313,7 @@ function data_boxing(data, r0, M = size(data, 2))
     prior, prior_perm = 1, permutations[1]
     # distributes all permutation indices into boxes
     for (index, perm) in enumerate(permutations)
-        if bins[perm] != bins[prior_perm]
+        if bins[perm] ≠ bins[prior_perm]
             push!(contents, permutations[prior:index-1])
             prior, prior_perm = index, perm
         end
@@ -279,13 +329,18 @@ For a vector of `boxes` and the indices of their `contents` inside of `data`,
 calculate the classic correlationsum of a radius or multiple radii `εs`.
 `w` is the Theiler window, for explanation see [`boxed_correlationsum`](@ref).
 """
-function boxed_correlationsum_2(boxes, contents, data, εs; w = 0)
+function boxed_correlationsum_2(boxes, contents, data, εs; w = 0, show_progress = false)
     Cs = zeros(eltype(data), length(εs))
     N = length(data)
-    for index in 1:length(boxes)
+    M = length(boxes)
+    if show_progress
+        progress = ProgressMeter.Progress(M; desc = "Boxed correlation sum: ", dt = 1.0)
+    end
+    for index in 1:M
         indices_neighbors = find_neighborboxes_2(index, boxes, contents)
         indices_box = contents[index]
         Cs .+= inner_correlationsum_2(indices_box, indices_neighbors, data, εs; w)
+        show_progress && ProgressMeter.update!(progress, index)
     end
     Cs .* (2 / ((N - w) * (N - w - 1)))
 end
@@ -352,15 +407,20 @@ For a vector of `boxes` and the indices of their `contents` inside of `data`,
 calculate the `q`-order correlationsum of a radius or radii `εs`.
 `w` is the Theiler window, for explanation see [`boxed_correlationsum`](@ref).
 """
-function boxed_correlationsum_q(boxes, contents, data, εs, q; w = 0)
+function boxed_correlationsum_q(boxes, contents, data, εs, q; w = 0, show_progress = false)
     q <= 1 && @warn "This function is currently not specialized for q <= 1" *
     " and may show unexpected behaviour for these values."
     Cs = zeros(eltype(data), length(εs))
     N = length(data)
-    for index in 1:length(boxes)
+    M = length(boxes)
+    if show_progress
+        progress = ProgressMeter.Progress(M; desc = "Boxed correlation sum: ", dt = 1.0)
+    end
+    for index in 1:M
         indices_neighbors = find_neighborboxes_q(index, boxes, contents, q)
         indices_box = contents[index]
         Cs .+= inner_correlationsum_q(indices_box, indices_neighbors, data, εs, q; w)
+        show_progress && ProgressMeter.update!(progress, index)
     end
     clamp.((Cs ./ ((N - 2w) * (N - 2w - 1) ^ (q-1))), 0, Inf) .^ (1 / (q-1))
 end
@@ -423,10 +483,15 @@ function inner_correlationsum_q(indices_X, indices_Y, data, εs, q::Real; norm =
     return Cs
 end
 
+#######################################################################################
+# Good boxsize estimates for boxed correlation sum
+#######################################################################################
 """
-    estimate_r0_theiler(data)
-Estimates a reasonable size for boxing the data before calculating the
-correlation dimension proposed by Theiler[^Theiler1987].
+    estimate_r0_theiler(X::Dataset) → r0, ε0
+Estimate a reasonable size for boxing the data `X` before calculating the
+[`boxed_correlationsum`](@ref) proposed by Theiler[^Theiler1987].
+Return the boxing size `r0` and minimum inter-point distance in `X`, `ε0`.
+
 To do so the dimension is estimated by running the algorithm by Grassberger and
 Procaccia[^Grassberger1983] with `√N` points where `N` is the number of total
 data points. Then the optimal boxsize ``r_0`` computes as
@@ -442,25 +507,33 @@ where ``R`` is the size of the chaotic attractor and ``\\nu`` is the estimated d
 function estimate_r0_theiler(data)
     N = length(data)
     mini, maxi = minmaxima(data)
-    R = maximum(maxi .- mini)
+    R = mean(maxi .- mini)
     # Sample √N datapoints for a rough estimate of the dimension.
     data_sample = data[unique(rand(1:N, ceil(Int, sqrt(N))))] |> Dataset
     # Define radii for the rough dimension estimate
-    lower = log10(minimum_pairwise_distance(data_sample)[1])
+    min_d, _ = minimum_pairwise_distance(data)
+    if min_d == 0
+        @warn(
+        "Minimum distance in the dataset is zero! Probably because of having data "*
+        "with low resolution, or duplicate data points. Setting to `d₊/1000` for now.")
+        min_d = R/(10^3)
+    end
+    lower = log10(min_d)
     εs = 10 .^ range(lower, stop = log10(R), length = 12)
     # Actually estimate the dimension.
     cm = correlationsum(data_sample, εs)
-    ν = linear_region(log.(εs), log.(cm), tol = 0.5)[2]
+    ν = linear_region(log.(εs), log.(cm), tol = 0.5, warning = false)[2]
     # The combination yields the optimal box size
     r0 = R * (2/N)^(1/ν)
+    return r0, min_d
 end
 
 """
-    estimate_r0_buenoorovio(X, M = size(X, 2))
+    estimate_r0_buenoorovio(X::Dataset, P = autoprismdim(X))
 Estimates a reasonable size for boxing the time series `X` proposed by
 Bueno-Orovio and Pérez-García[^Bueno2007] before calculating the correlation
 dimension as presented by Theiler[^Theiler1983]. If instead of boxes, prisms
-are chosen everything stays the same but `M` is the dimension of the prism.
+are chosen everything stays the same but `P` is the dimension of the prism.
 To do so the dimension `ν` is estimated by running the algorithm by Grassberger
 and Procaccia[^Grassberger1983] with `√N` points where `N` is the number of
 total data points.
@@ -473,9 +546,9 @@ boxes `η_ℓ`.
 The optimal number of filled boxes `η_opt` is calculated by minimising the number
 of calculations.
 ```math
-\\eta_\\textrm{opt} = N^{2/3}\\cdot \\frac{3^\\nu - 1}{3^M - 1}^{1/2}.
+\\eta_\\textrm{opt} = N^{2/3}\\cdot \\frac{3^\\nu - 1}{3^P - 1}^{1/2}.
 ```
-`M` is the dimension of the data or the number of edges on the prism that don't
+`P` is the dimension of the data or the number of edges on the prism that don't
 span the whole dataset.
 
 Then the optimal boxsize ``r_0`` computes as
@@ -489,33 +562,40 @@ r_0 = \\ell / \\eta_\\textrm{opt}^{1/\\nu}.
 
 [^Grassberger1983]: Grassberger and Proccacia, [Characterization of strange attractors, PRL 50 (1983)](https://journals-aps-org.e-bis.mpimet.mpg.de/prl/abstract/10.1103/PhysRevLett.50.346)
 """
-function estimate_r0_buenoorovio(X, M = size(X, 2))
+function estimate_r0_buenoorovio(X, P = autoprismdim(X))
     mini, maxi = minmaxima(X)
     N = length(X)
-    R = maximum(maxi .- mini)
+    R = mean(maxi .- mini)
     # The possibility of a bad pick exists, if so, the calculation is repeated.
     ν = zero(eltype(X))
+    min_d, _ = minimum_pairwise_distance(X)
+    if min_d == 0
+        @warn(
+        "Minimum distance in the dataset is zero! Probably because of having data "*
+        "with low resolution, or duplicate data points. Setting to `d₊/1000` for now.")
+        min_d = R/(10^3)
+    end
+    
     # Sample N/10 datapoints out of data for rough estimate of effective size.
     sample1 = X[unique(rand(1:N, N÷10))] |> Dataset
     r_ℓ = R / 10
-    η_ℓ = length(data_boxing(sample1, r_ℓ)[1])
+    η_ℓ = length(data_boxing(sample1, r_ℓ, P)[1])
     r0 = zero(eltype(X))
     while true
         # Sample √N datapoints for rough dimension estimate
         sample2 = X[unique(rand(1:N, ceil(Int, sqrt(N))))] |> Dataset
         # Define logarithmic series of radii.
-        lower = log10(minimum_pairwise_distance(X)[1])
-        εs = 10 .^ range(lower, stop = log10(R), length = 16)
+        εs = 10.0 .^ range(log10(min_d), log10(R); length = 16)
         # Estimate ν from a sample using the Grassberger Procaccia algorithm.
         cm = correlationsum(sample2, εs)
-        ν = linear_region(log.(εs), log.(cm), tol = 0.5)[2]
+        ν = linear_region(log.(εs), log.(cm); tol = 0.5, warning = false)[2]
         # Estimate the effictive size of the chaotic attractor.
         ℓ = r_ℓ * η_ℓ^(1/ν)
         # Calculate the optimal number of filled boxes according to Bueno-Orovio
-        η_opt = N^(2/3) * ((3^ν - 1/2) / (3^M - 1))^(1/2)
+        η_opt = N^(2/3) * ((3^ν - 1/2) / (3^P - 1))^(1/2)
         # The optimal box size is the effictive size divided by the box number # to the power of the inverse dimension.
         r0 = ℓ / η_opt^(1/ν)
         !isnan(r0) && break
     end
-    return r0
+    return r0, min_d
 end