+++ test/chaosdetection/partially_predictable_tests.jl
@@ -1,7 +1,5 @@
 using ChaosTools
 using Test
-using OrdinaryDiffEq: Vern9
-lorenz = Systems.lorenz
 using Random
 
 ν_thresh_lower, ν_thresh_upper = 0.1, 0.9
@@ -12,9 +10,9 @@ println("\nTesting Partially predictable chaos...")
 @testset "Predictability Lorenz" begin
     @testset "strongly chaotic" begin
         Random.seed!(12)
-        lz = lorenz(ρ=180.70)
+        lz = Systems.lorenz(ρ=180.70)
         @time chaos_type, ν, C = predictability(lz;
-            λ_max = 1.22, alg=Vern9(), maxiters=1e9, T_max = 1e3, n_samples = 100,
+            λ_max = 1.22, maxiters=1e9, T_max = 1e3, n_samples = 100,
         )
         @test chaos_type == :SC
         @test ν < ν_thresh_lower
@@ -23,9 +21,9 @@ println("\nTesting Partially predictable chaos...")
     end
     @testset "PPC 1" begin
         Random.seed!(12)
-        lz = lorenz(ρ=180.78)
+        lz = Systems.lorenz(ρ=180.78)
         @time chaos_type, ν, C = predictability(lz;
-            λ_max = 0.4, alg=Vern9(), maxiters=1e9, n_samples = 100, T_max = 400
+            λ_max = 0.4, maxiters=1e9, n_samples = 100, T_max = 400
         )
         @test chaos_type == :PPC
         @test ν < ν_thresh_lower
@@ -34,9 +32,9 @@ println("\nTesting Partially predictable chaos...")
     end
     @testset "PPC 2" begin
         Random.seed!(12)
-        lz = lorenz(ρ=180.95)
+        lz = Systems.lorenz(ρ=180.95)
         @time chaos_type, ν, C = predictability(lz;
-            λ_max = 0.1, alg=Vern9(), maxiters=1e9, n_samples = 100, T_max = 400
+            λ_max = 0.1, maxiters=1e9, n_samples = 100, T_max = 400
         )
         @test chaos_type == :PPC
         @test ν < ν_thresh_lower
@@ -45,9 +43,9 @@ println("\nTesting Partially predictable chaos...")
     end
     @testset "laminar" begin
         Random.seed!(12)
-        lz = lorenz(ρ=181.10)
+        lz = Systems.lorenz(ρ=181.10)
         @time chaos_type, ν, C = predictability(lz;
-            λ_max = 0.01, T_max = 400, alg=Vern9(), n_samples = 100, maxiters=1e9
+            λ_max = 0.01, T_max = 400, n_samples = 100, maxiters=1e9
         )
         @test chaos_type == :REG
         @test ν > ν_thresh_upper